Update `rand` and `chrono` dependencies
The latest version of `rand` is `0.8`, which updates a number of its own
dependencies to the latest versions.

The `chrono` crate is still at 0.4 so this isn't an upgrade per se,
however the default feature `oldclock` pulls in the very old `time` 0.1
crate, and for ULID generation none of the legacy `time` 0.1
functionality is required.